changed the yesfsdir...
to an environment variable.
chaned FILECOUNT to an environment
variable
included func.regex

Invoked parallel execution of the yes1 script to convert an individual
file to a hashed object.  We may have to add -j to increase the
parallelism of this to increase CPU and I/O utilization while running
these scripts.